pyproject.toml
[project]
name = "gemini-interactions-mcp"
version = "0.1.0"
description = "FastMCP server for Gemini API - provides web search and grounded AI answers using Google Search"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/gemini-interactions-mcp"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
